backend/lfras_phone/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg[binary]==3.1.19
python-dotenv==1.0.1
boto3==1.34.145
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
drf-spectacular==0.27.2